refactor: 컴포넌트 섹션 분리
