+++ src/components/SocialWorker/RegisterMatchingElder/TitleInputSection.tsx
@@ -1,0 +1,52 @@
+import { PlainInputBox } from '@/components/common/InputBox/PlainInputBox';
+import styled from 'styled-components';
+
+interface Props {
+  title: string;
+  setTitle: (value: string) => void;
+}
+
+export const TitleInputSection = ({ title, setTitle }: Props) => {
+  return (
+    <TitleContainer>
+      <div className="name">
+        <span>제목</span>
+        <span className="highlight">*</span>
+      </div>
+      <PlainInputBox
+        state="default"
+        placeholder="공고 제목을 입력하세요."
+        value={title}
+        onChange={(e) => setTitle(e.target.value)}
+        width="100%"
+        guide=""
+      />
+    </TitleContainer>
+  );
+};
+
+const TitleContainer = styled.div`
+  display: flex;
+  padding: 40px 20px 0px 20px;
+  align-items: flex-start;
+  width: 100%;
+  gap: 8px;
+
+  flex-direction: column;
+
+  font-weight: ${({ theme }) => theme.typography.fontWeight.body2};
+  font-weight: ${({ theme }) => theme.typography.fontWeight.semibold};
+  color: ${({ theme }) => theme.colors.gray900};
+
+  .highlight {
+    font-weight: ${({ theme }) => theme.typography.fontWeight.body2};
+    font-weight: ${({ theme }) => theme.typography.fontWeight.semibold};
+    color: ${({ theme }) => theme.colors.mainBlue};
+  }
+
+  .name {
+    display: flex;
+    flex-direction: row;
+    gap: 2px;
+  }
+`;
